pyproject.toml
[project]
name = "bible-app"
version = "0.1.0"
description = "Bible Search Application"
dependencies = [
    "taipy-gui>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"